Adjust annotations of serrors
Previously we required all elements of `errCtx` to be low. Aside from
the fact that this annotation didn't make sense in case `errCtx`
contained values of reference types, it was also unnecessary, since (by
now) we only require that whether the resulting error is not nil to be
low -- which was already guaranteed by the contract.

We do, however, need to ensure that there is no non-low control flow.
This might require further annotations in the future.